Abstract parts of chado install into generic bulk postgresql schema installer.
