+++ tripal/tripal.services.yml
@@ -3,4 +3,6 @@ services:
     class: Drupal\tripal\Access\CustomAccessCheck
     arguments: ['@current_user']
     tags:
-      - { name: access_check, applies_to: _tripal_access_check } 
+      - { name: access_check, applies_to: _tripal_access_check }
+  tripal.bulkPgSchemaInstaller:
+      class: Drupal\tripal\Services\bulkPgSchemaInstaller
